Separate all types in this to their own files
