+++ Terminal.Gui/Drawing/Color.cs
@@ -1,5 +1,4 @@
 #nullable enable
-using System.Buffers.Binary;
 using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using System.Diagnostics.Contracts;
@@ -11,188 +10,6 @@ using System.Text;
 using System.Text.Json.Serialization;
 
 namespace Terminal.Gui;
-
-/// <summary>
-/// Defines the 16 legacy color names and values that can be used to set the
-/// foreground and background colors in Terminal.Gui apps. Used with <see cref="Color"/>.
-/// </summary>
-/// <remarks>
-///         <para>
-///         These colors match the 16 colors defined for ANSI escape sequences for 4-bit (16) colors.
-///         </para>
-///         <para>
-///         For terminals that support 24-bit color (TrueColor), the RGB values for each of these colors can be configured
-///         using the
-///         <see cref="Color.Colors"/> property.
-///         </para>
-/// </remarks>
-public enum ColorName {
-	/// <summary>
-	/// The black color. ANSI escape sequence: <c>\u001b[30m</c>.
-	/// </summary>
-	Black,
-
-	/// <summary>
-	/// The blue color. ANSI escape sequence: <c>\u001b[34m</c>.
-	/// </summary>
-	Blue,
-
-	/// <summary>
-	/// The green color. ANSI escape sequence: <c>\u001b[32m</c>.
-	/// </summary>
-	Green,
-
-	/// <summary>
-	/// The cyan color. ANSI escape sequence: <c>\u001b[36m</c>.
-	/// </summary>
-	Cyan,
-
-	/// <summary>
-	/// The red color. ANSI escape sequence: <c>\u001b[31m</c>.
-	/// </summary>
-	Red,
-
-	/// <summary>
-	/// The magenta color. ANSI escape sequence: <c>\u001b[35m</c>.
-	/// </summary>
-	Magenta,
-
-	/// <summary>
-	/// The yellow color (also known as Brown). ANSI escape sequence: <c>\u001b[33m</c>.
-	/// </summary>
-	Yellow,
-
-	/// <summary>
-	/// The gray color (also known as White). ANSI escape sequence: <c>\u001b[37m</c>.
-	/// </summary>
-	Gray,
-
-	/// <summary>
-	/// The dark gray color (also known as Bright Black). ANSI escape sequence: <c>\u001b[30;1m</c>.
-	/// </summary>
-	DarkGray,
-
-	/// <summary>
-	/// The bright blue color. ANSI escape sequence: <c>\u001b[34;1m</c>.
-	/// </summary>
-	BrightBlue,
-
-	/// <summary>
-	/// The bright green color. ANSI escape sequence: <c>\u001b[32;1m</c>.
-	/// </summary>
-	BrightGreen,
-
-	/// <summary>
-	/// The bright cyan color. ANSI escape sequence: <c>\u001b[36;1m</c>.
-	/// </summary>
-	BrightCyan,
-
-	/// <summary>
-	/// The bright red color. ANSI escape sequence: <c>\u001b[31;1m</c>.
-	/// </summary>
-	BrightRed,
-
-	/// <summary>
-	/// The bright magenta color. ANSI escape sequence: <c>\u001b[35;1m</c>.
-	/// </summary>
-	BrightMagenta,
-
-	/// <summary>
-	/// The bright yellow color. ANSI escape sequence: <c>\u001b[33;1m</c>.
-	/// </summary>
-	BrightYellow,
-
-	/// <summary>
-	/// The White color (also known as Bright White). ANSI escape sequence: <c>\u001b[37;1m</c>.
-	/// </summary>
-	White
-}
-
-/// <summary>
-/// The 16 foreground color codes used by ANSI Esc sequences for 256 color terminals. Add 10 to these values for background
-/// color.
-/// </summary>
-public enum AnsiColorCode {
-	/// <summary>
-	/// The ANSI color code for Black.
-	/// </summary>
-	BLACK = 30,
-
-	/// <summary>
-	/// The ANSI color code for Red.
-	/// </summary>
-	RED = 31,
-
-	/// <summary>
-	/// The ANSI color code for Green.
-	/// </summary>
-	GREEN = 32,
-
-	/// <summary>
-	/// The ANSI color code for Yellow.
-	/// </summary>
-	YELLOW = 33,
-
-	/// <summary>
-	/// The ANSI color code for Blue.
-	/// </summary>
-	BLUE = 34,
-
-	/// <summary>
-	/// The ANSI color code for Magenta.
-	/// </summary>
-	MAGENTA = 35,
-
-	/// <summary>
-	/// The ANSI color code for Cyan.
-	/// </summary>
-	CYAN = 36,
-
-	/// <summary>
-	/// The ANSI color code for White.
-	/// </summary>
-	WHITE = 37,
-
-	/// <summary>
-	/// The ANSI color code for Bright Black.
-	/// </summary>
-	BRIGHT_BLACK = 90,
-
-	/// <summary>
-	/// The ANSI color code for Bright Red.
-	/// </summary>
-	BRIGHT_RED = 91,
-
-	/// <summary>
-	/// The ANSI color code for Bright Green.
-	/// </summary>
-	BRIGHT_GREEN = 92,
-
-	/// <summary>
-	/// The ANSI color code for Bright Yellow.
-	/// </summary>
-	BRIGHT_YELLOW = 93,
-
-	/// <summary>
-	/// The ANSI color code for Bright Blue.
-	/// </summary>
-	BRIGHT_BLUE = 94,
-
-	/// <summary>
-	/// The ANSI color code for Bright Magenta.
-	/// </summary>
-	BRIGHT_MAGENTA = 95,
-
-	/// <summary>
-	/// The ANSI color code for Bright Cyan.
-	/// </summary>
-	BRIGHT_CYAN = 96,
-
-	/// <summary>
-	/// The ANSI color code for Bright White.
-	/// </summary>
-	BRIGHT_WHITE = 97
-}
 
 /// <summary>
 /// Represents a 24-bit color. Provides automatic mapping between the legacy 4-bit (16 color) system and 24-bit colors (see
